account components refactores, also fixed suapbase error on getting uuid which was undefined
